pyproject.toml
[project]
name = "pyscigl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"


dependencies = [
    "glfw>=2.10.0",
    "numpy>=2.4.1",
    "pathlib>=1.0.1",
    "pillow>=12.1.1",
    "pyopengl>=3.1.10",
    "pyopengl-accelerate>=3.1.10",
    "typer>=0.22.0",
]

[tool.uv]
package = true

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project.scripts]
pyscigl = "pyscigl.__main__:app"

